Server/requirements2.txt
absl-py==0.8.0
alabaster==0.7.12
amqp==2.4.2
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
aniso8601==3.0.2
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
audioread==2.1.8
autopep8==1.4.3
Babel==2.6.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.6
beautifulsoup4==4.8.0
billiard==3.6.0.0
bitarray==1.0.1
bkcharts==0.2
bleach==2.1.4
bokeh==1.3.4
boto==2.49.0
boto3==1.9.130
botocore==1.12.130
Bottleneck==1.2.1
braintree==3.49.0
cairocffi==1.0.2
CairoSVG==2.3.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
conda==4.7.12
conda-build==3.18.9
conda-package-handling==1.5.0
conda-verify==3.4.2
config==0.4.0
contextlib2==0.5.5
cryptography==2.7
cssselect2==0.2.1
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.3.0
decorator==4.4.0
defusedxml==0.5.0
dill==0.3.1.1
distributed==2.3.2
dj-database-url==0.5.0
dj-email-url==0.2.0
Django==2.2
django-appconf==1.0.3
django-babel==0.6.2
django-bootstrap4==0.0.8
django-cache-url==3.0.0
django-celery-results==1.0.4
django-countries==5.3.3
django-debug-toolbar==1.11
django-debug-toolbar-request-history==0.0.9
django-elasticsearch-dsl==0.5.1
django-filter==2.1.0
django-fsm==2.6.0
django-graphql-jwt==0.2.1
django-heroku==0.3.1
django-impersonate==1.4
django-js-asset==1.2.2
django-measurement==3.0.0
django-mptt==0.9.1
django-phonenumber-field==2.3.1
django-prices==1.0.4
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.2
django-recaptcha==1.5.0
django-redis==4.10.0
django-render-block==0.5
django-silk==2.0.0
django-storages==1.7.1
django-templated-email==2.3.0
django-versatileimagefield==1.10
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
docutils==0.14
elasticsearch==6.3.1
elasticsearch-dsl==6.0.1
entrypoints==0.3
et-xmlfile==1.0.1
Faker==1.0.4
fastcache==1.1.0
filelock==3.0.12
Flask==1.0.2
Flask-Gravatar==0.5.0
Flask-SQLAlchemy==2.3.2
freezegun==0.3.11
fsspec==0.4.0
future==0.17.1
gast==0.3.1
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
google-i18n-address==2.3.4
google-measurement-protocol==1.0.0
google-pasta==0.1.7
gprof2dot==2016.10.13
graphene==2.1.3
graphene-django==2.2.0
graphene-django-optimizer==0.3.6
graphql-core==2.1
graphql-relay==0.4.5
greenlet==0.4.15
grpcio==1.23.0
gunicorn==20.0.4
h5py==2.10.0
heapdict==1.0.0
hickle==3.4.5
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.19
imutils==0.5.3
inflection==0.4.0
ipaddress==1.0.22
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.17
itsdangerous==1.1.0
japronto==0.1.2
jdcal==1.4.1
jedi==0.15.1
jeepney==0.4
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
json5==0.8.5
jsonfield==2.0.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.3
jupyterlab-server==1.0.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.1.0
kombu==4.5.0
lazy-object-proxy==1.3.1
libarchive-c==2.8
librosa==0.7.0
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
maxminddb==1.4.1
maxminddb-geolite2==2018.703
mccabe==0.6.1
measurement==2.0.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nlpaug==0.0.10
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.17.2
numpydoc==0.9.1
oauthlib==3.0.1
olefile==0.46
opencv-python==4.1.0.25
openpyxl==2.6.2
packaging==19.1
pandas==0.24.2
pandocfilters==1.4.2
paramiko==2.5.0
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.4
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pgadmin4==4.8
phonenumberslite==8.10.9
pickleshare==0.7.5
Pillow==5.4.1
pkginfo==1.5.0.1
pluggy==0.12.0
ply==3.11
prices==1.0.0
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psutil==5.5.1
psycopg2==2.8.5
psycopg2-binary==2.8.1
ptyprocess==0.6.0
purl==1.5
py==1.8.0
PyAudio==0.2.11
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pyflakes==2.1.1
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.3.1
PyNaCl==1.3.0
pynput==1.4.2
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
Pyphen==0.9.5
pyrsistent==0.14.11
PySocks==1.7.0
pytest==5.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.2
python-dateutil==2.8.0
python-speech-features==0.6
python-xlib==0.25
python3-openid==3.1.0
pytz==2018.9
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.5.7
qtconsole==4.5.5
QtPy==1.9.0
raven==6.9.0
razorpay==1.2.0
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
resampy==0.2.2
rope==0.14.0
ruamel-yaml==0.15.46
Rx==1.6.1
s3transfer==0.2.0
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.1
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snowballstemmer==1.9.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
sounddevice==0.3.14
SoundFile==0.10.2
soupsieve==1.9.2
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.1
SQLAlchemy==1.3.7
sqlparse==0.2.4
sshtunnel==0.1.5
statsmodels==0.10.1
stripe==2.24.1
sympy==1.3
tables==3.5.2
tblib==1.4.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
Theano==1.0.4
tinycss2==1.0.2
toolz==0.10.0
tornado==6.0.3
tqdm==4.32.1
traitlets==4.3.2
typed-ast==1.3.4
typing==3.6.6
unicodecsv==0.14.1
urllib3==1.24.1
uvloop==0.12.2
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.7
WeasyPrint==46
webencodings==0.5.1
Werkzeug==0.15.6
whitenoise==5.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
zict==1.0.0
zipp==0.5.2
